cleanup & small optimisations

+++ firmware_17xx/synth/utils.c
@@ -22,7 +22,7 @@ inline uint16_t satAddU16S32(uint16_t a, int32_t b)
 	r+=b;
 	r=__USAT(r,16);
 	
-	return (uint16_t)r;
+	return r;
 }
 
 inline uint16_t satAddU16S16(uint16_t a, int16_t b)
